Refactoring and small bug fixes
git-svn-id: https://svn.code.sf.net/p/paintbrush/code/Paintbrush2@11 c246e104-b051-0410-8ce7-bd27a59b7068

+++ SWAirbrushTool.m
@@ -35,7 +35,7 @@
 	[path setLineCapStyle:NSRoundLineCapStyle];
 	NSBezierPath *circle = [NSBezierPath bezierPathWithOvalInRect:redrawRect];
 	
-	int i, x, y;
+	NSInteger i, x, y;
 	for (i = 0; i < (lineWidth*lineWidth)/2; i++) {
 		do {
 			x = (random() % (4*(int)lineWidth))+end.x - 2*lineWidth;
